✨ Added AzulAi and additional features to support it
So far AzulAi is capable of determining all valid moves and, when asked, it just picks a random valid move.

@TODO
Rarely a weird bug occurs where the last tile cannot be picked. Somehow its color code is the Game Stone (which is already in some players basement) but is displayed as another color

@TODO
The TileBag needs to be refilled when empty